TASK 4
Analyzing Sequence Composition (Nucleotide or Amino Acid Frequencies) 
Objective: Analyze the nucleotide composition of the TNF gene sequence.
The frequencies of the nucleotides present in the gene sequence have been displayed with several information such as the length of the base pairs, the molecular weight of single and double stranded DNA molecule, GC and AT percentage ratio as well as the number of times each nucleotides appear. This information is crucial in drug discovery, and protein analysis. 
It is evident in the graph that there is a higher GC content in the Human TNF gene sequence which means that there is a higher percentage of coding regions in the sequence.