Searching for Functional Motifs in the Human TNF Genome or Transcriptome Using MEME Suite  Identifying these motifs within a genome or transcriptome is crucial for understanding gene function, regulation, and protein structure. Objective: Search for functional motifs in the TNF gene sequence using MEME Suite.
Meme suite came up with three motifs as instructed for the Human TNF gene sequence. Information such as the E-value, and the number of the binding sites for each motif (there are represented in colored bars under the motifs location tab) are displayed. For further information on each motif, the more icon is clicked, and the result is displayed below. It gave more information on the type of strand present, where the motif occurrence starts and other pertinent information.